Updated github code to match code in Plastic SCM after issue #33 completed
