+++ Code/TanukiTeleporterRegion.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3fa44987b62c6314abf4d8abd7c3a1c5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
