+++ Code/TanukiTeleporterRegion.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 33b0f16d16d270744ba66d279d846510
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
