+++ Editor/Tanuki-Documentation.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e5d36a8bcecef7947840f0580e4ff609
+guid: ba64aaed3708c4a22bf08ff32888aa97
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
