+++ Editor/Tanuki-Main.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 52d02aa6326b547cc998099a6e334980
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
