+++ Editor/Tanuki-Tools.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8e8b605a29c9c044ea42d744d5fca572
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
