+++ Editor/Tanuki-Whatever.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5bba3cae867ff403cb57d6d4fe78d87d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
